UI: Fix projector connecting the wrong source signal
The signal name is "remove" for when a source is removed, not "removed".

This is proof that I should never have relied on strings for signals.
The original intention for string-based signals was to make them
programmable and scriptable, but honestly that use-case (that never
happened and will likely never happen) was foolish to program around.
These should have been fixed macros from the beginning.